feat: use external standalone `gltf-type-toolkit` plugin (#140)
