+++ package.json
@@ -37,7 +37,7 @@
     "build-only": "vite build",
     "test:types": "vue-tsc --build --force",
     "postinstall": "run-s \"generate:gltf-models\"",
-    "generate:gltf-models": "tsx tools/generate-model-types.ts",
+    "generate:gltf-models": "gltf-codegen",
     "lint": "eslint .",
     "lint:fix": "run-s \"lint --fix\"",
     "lint:print-current-config": "mkdir -p ./dist/_eslintConfig && eslint --print-config file.json > ./dist/_eslintConfig/json.json && eslint --print-config file.vue > ./dist/_eslintConfig/vue.json && eslint --print-config file.js > ./dist/_eslintConfig/js.json && eslint --print-config file.ts > ./dist/_eslintConfig/ts.json && eslint --print-config file.md > ./dist/_eslintConfig/md.json && eslint --print-config file.html > ./dist/_eslintConfig/html.json && eslint --print-config file.yaml > ./dist/_eslintConfig/yaml.json && eslint --print-config file.xml > ./dist/_eslintConfig/xml.json && eslint --print-config file.toml > ./dist/_eslintConfig/toml.json && eslint --print-config file.gql > ./dist/_eslintConfig/gql.json && eslint --print-config file.css > ./dist/_eslintConfig/css.json && eslint --print-config file.astro > ./dist/_eslintConfig/astro.json && eslint --print-config file.svelte > ./dist/_eslintConfig/svelte.json && eslint --print-config file.less > ./dist/_eslintConfig/less.json && eslint --print-config file.scss > ./dist/_eslintConfig/scss.json && eslint --print-config file.pcss > ./dist/_eslintConfig/pcss.json"
@@ -50,6 +50,7 @@
     "@iconify-json/ph": "~1.2.1",
     "@iconify/tailwind": "~1.1.3",
     "@rushstack/eslint-patch": "~1.10.4",
+    "@todde.tv/gltf-type-toolkit": "~1.1.0",
     "@tresjs/cientos": "~4.0.2",
     "@tresjs/core": "~4.2.10",
     "@tsconfig/node20": "~20.1.4",
