+++ src/components/models/Island.vue
@@ -1,9 +1,12 @@
 <script setup lang="ts">
-import modelLoader from '@/assets/models/Island/Island.gltf'
+import { getNode, IslandScene } from '@/assets/models/Island/Island.gltf.js'
 import { addShadowAndAddToGroup } from '@/utils/threeHelper.js'
 import { shallowRef, watch } from 'vue'
 
-const { scenes } = await modelLoader
+const model = await getNode(IslandScene)
+const seabedPlane = await getNode(IslandScene.zligislandlvl0_seabed001)
+seabedPlane.receiveShadow = false
+seabedPlane.castShadow = false
 
 const groupWrapperRef = shallowRef()
 
@@ -12,16 +15,7 @@ watch(groupWrapperRef, (newValue) => {
     return
   }
 
-  const model = scenes.Island.Scene.clone()
   addShadowAndAddToGroup(newValue, model, 'both')
-
-  // TODO do not use `getObjectByName` bc a string is not catched when name changes
-  const seabedPlane = model.getObjectByName('zligislandlvl0_seabed001')
-  if (!seabedPlane) {
-    return
-  }
-  seabedPlane.receiveShadow = false
-  seabedPlane.castShadow = false
 })
 </script>
