+++ src/components/models/Waterfall.vue
@@ -1,5 +1,5 @@
 <script setup lang="ts">
-import modelLoader from '@/assets/models/Waterfall/Waterfall.gltf'
+import { getNode, WaterfallScene } from '@/assets/models/Waterfall/Waterfall.gltf.js'
 import { useDebugStore } from '@/composables/useDebugStore.js'
 import { useGameTimeStore } from '@/composables/useGameTimeStore.js'
 import { addShadowAndAddToGroup } from '@/utils/threeHelper.js'
@@ -7,7 +7,6 @@ import { getWaterMaterial } from '@/utils/WaterShader.js'
 import { storeToRefs } from 'pinia'
 import { shallowRef, watch } from 'vue'
 
-const { scenes } = await modelLoader
 const { onTick } = useGameTimeStore()
 const { showWaterWireframe } = storeToRefs(useDebugStore())
 
@@ -27,6 +26,15 @@ const waterMaterialDown = getWaterMaterial({
   waveTangentialAmplitude: 0.0,
 })
 
+const model1 = await getNode(WaterfallScene.zligwaterfalllvl0001)
+model1.material = waterMaterial
+
+const model2 = await getNode(WaterfallScene.zligwaterfalllvl1002)
+model2.material = waterMaterial
+
+const model3 = await getNode(WaterfallScene.zligwaterfalllvl1_to_lvl0001)
+model3.material = waterMaterialDown
+
 onTick(({ ambientAnimationDelta }) => {
   waterMaterial.uniforms.time.value += ambientAnimationDelta
   waterMaterialDown.uniforms.time.value += ambientAnimationDelta
@@ -43,16 +51,8 @@ watch(groupWrapperRef, (newValue) => {
     return
   }
 
-  const model1 = scenes.Waterfall.Object.zligwaterfalllvl0001.clone()
-  model1.material = waterMaterial
   addShadowAndAddToGroup(newValue, model1, 'receive')
-
-  const model2 = scenes.Waterfall.Object.zligwaterfalllvl1002.clone()
-  model2.material = waterMaterial
   addShadowAndAddToGroup(newValue, model2, 'receive')
-
-  const model3 = scenes.Waterfall.Object.zligwaterfalllvl1_to_lvl0001.clone()
-  model3.material = waterMaterialDown
   addShadowAndAddToGroup(newValue, model3, 'receive')
 })
 </script>
