+++ src/composables/useBuildingAreasStore.ts
@@ -1,4 +1,4 @@
-import modelLoader from '@/assets/models/Island/Island.gltf'
+import { getNode, IslandScene } from '@/assets/models/Island/Island.gltf.js'
 import { defineStore } from 'pinia'
 import { readonly, ref } from 'vue'
 import type { BuildingAreaId } from '@/game-logic/types.js'
@@ -12,9 +12,7 @@ export const useBuildingAreasStore = defineStore('building-areas', () => {
   }[]>([])
 
   async function init() {
-    const { scenes: { Island } } = await modelLoader
-
-    Island.Scene.traverse((obj) => {
+    (await getNode(IslandScene)).traverse((obj) => {
       if (obj.userData.isBuildArea && obj.name) {
         BuildingAreas.value.push({
           id: obj.name,
