+++ tsconfig.app.json
@@ -6,17 +6,13 @@
 
     "baseUrl": ".",
     "paths": {
-      "@/*": [
-        "./src/*",
-        "./node_modules/.tmp/model-types/src/*"
-      ]
+      "@/*": ["./src/*"]
     }
   },
   "include": [
     "env.d.ts",
     "src/**/*",
-    "src/**/*.vue",
-    "node_modules/.tmp/model-types/**/*.d.ts"
+    "src/**/*.vue"
   ],
   "exclude": [
     "src/**/__tests__/*"
