+++ vite.config.ts
@@ -1,9 +1,10 @@
 import { fileURLToPath, URL } from 'node:url'
+// eslint-disable-next-line import/extensions
+import gltf from '@todde.tv/gltf-type-toolkit/vite'
 import { templateCompilerOptions } from '@tresjs/core'
 import vue from '@vitejs/plugin-vue'
 import { defineConfig } from 'vite'
 import vueDevTools from 'vite-plugin-vue-devtools'
-import { gltf } from './tools/vite-plugin-gltf.js'
 
 // https://vitejs.dev/config/
 export default defineConfig({
@@ -12,7 +13,10 @@ export default defineConfig({
       ...templateCompilerOptions,
     }),
     vueDevTools(),
-    gltf(),
+    gltf({
+      customGltfLoaderModule: '@/utils/gltfLoader.ts',
+      verbose: true,
+    }),
   ],
   resolve: {
     alias: {
